Practice RubyDoctest
It is a simple framework that emulates IRB with the simple-prompt.
Fast and easy to test things out, even without running IRB but with the
advantage of recording our IRB session in the code with what is being
tested.  We probably won't use it all the time, perhaps we will use it
in an external file sometimes.  Still, the simplets thing possible to
teach testing, I think.  Well, except for using puts but then you have
to remember to comment those out, or delete them.  Yuck!